pyproject.toml
[project]
name = "llm-automation-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.115.8",
    "numpy>=2.2.3",
    "openai>=1.63.0",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=3.4.1",
    "uvicorn>=0.34.0",
]
